rpc: factor out health check logic
