+++ crates/sui-rpc-api/src/service/health.rs
@@ -1,0 +1,37 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::time::Duration;
+use std::time::SystemTime;
+
+use crate::Result;
+use crate::RpcService;
+
+impl RpcService {
+    /// Perform a simple health check on the service.
+    ///
+    /// The threshold, or delta, between the server's system time and the timestamp in the most
+    /// recently executed checkpoint for which the server is considered to be healthy.
+    ///
+    /// If not provided, the server will be considered healthy if it can simply fetch the latest
+    /// checkpoint from its store.
+    pub fn health_check(&self, threshold_seconds: Option<u32>) -> Result<()> {
+        let summary = self.reader.inner().get_latest_checkpoint()?;
+
+        // If we have a provided threshold, check that it's close to the current time
+        if let Some(threshold_seconds) = threshold_seconds {
+            let latest_chain_time = summary.timestamp();
+
+            let threshold = SystemTime::now() - Duration::from_secs(threshold_seconds as u64);
+
+            if latest_chain_time < threshold {
+                return Err(anyhow::anyhow!(
+                    "The latest checkpoint timestamp is less than the provided threshold"
+                )
+                .into());
+            }
+        }
+
+        Ok(())
+    }
+}
