+++ crates/sui-rpc-api/src/service/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+mod health;
 mod info;
